restructured footer
