+++ app/(default)/layout.tsx
@@ -6,7 +6,7 @@ import AOS from "aos";
 import "aos/dist/aos.css";
 
 import PageIllustration from "@/components/page-illustration";
-import Footer from "@/components/ui/footer";
+import Footer from "@/components/footer";
 
 export default function DefaultLayout({
   children,
